+++ Mac/localDebugInstall.command
@@ -13,7 +13,6 @@ export UXPS_COMMANDS_DIR=`pwd`/
 
 . setTarget.command
 
-
 echo ""
 echo "Removing ${TARGET_SCRIPT_ROOT_DIR}" 
 echo ""
@@ -30,8 +29,22 @@ echo ""
 echo "Creating temporary symbolic links to the script and its 'helpers' directory inside directory ${TARGET_SCRIPT_ROOT_DIR}"
 echo ""
 
-ln -s "${PROJECT_ROOT_DIR}run_as_UXPScript.${TARGET_FILENAME_EXTENSION}" "${TARGET_SCRIPT_ROOT_DIR}run_as_UXPScript.${TARGET_FILENAME_EXTENSION}"
-ln -s "${PROJECT_ROOT_DIR}run_as_ExtendScript.jsx" "${TARGET_SCRIPT_ROOT_DIR}run_as_ExtendScript.jsx"
+if [ -f "${PROJECT_ROOT_DIR}run_as_UXPScript.${TARGET_FILENAME_EXTENSION}" ]; then
+    ln -s "${PROJECT_ROOT_DIR}run_as_UXPScript.${TARGET_FILENAME_EXTENSION}" "${TARGET_SCRIPT_ROOT_DIR}run_as_UXPScript.${TARGET_FILENAME_EXTENSION}"
+fi
+
+if [ -f "${PROJECT_ROOT_DIR}run.${TARGET_FILENAME_EXTENSION}" ]; then
+    ln -s "${PROJECT_ROOT_DIR}run.${TARGET_FILENAME_EXTENSION}" "${TARGET_SCRIPT_ROOT_DIR}run.${TARGET_FILENAME_EXTENSION}"
+fi
+
+if [ -f "${PROJECT_ROOT_DIR}run_as_ExtendScript.jsx" ]; then
+    ln -s "${PROJECT_ROOT_DIR}run_as_ExtendScript.jsx" "${TARGET_SCRIPT_ROOT_DIR}run_as_ExtendScript.jsx"
+fi
+
+if [ -f "${PROJECT_ROOT_DIR}run.jsx" ]; then
+    ln -s "${PROJECT_ROOT_DIR}run.jsx" "${TARGET_SCRIPT_ROOT_DIR}run.jsx"
+fi
+
 ln -s "${PROJECT_ROOT_DIR}${DESPACED_TARGET_NAME}.js" "${TARGET_SCRIPT_ROOT_DIR}${DESPACED_TARGET_NAME}.js"
 ln -s "${PROJECT_ROOT_DIR}${DESPACED_TARGET_NAME}_helpers" "${TARGET_SCRIPT_ROOT_DIR}${DESPACED_TARGET_NAME}_helpers"
